+++ easy_server/_srvdef.py
@@ -11,7 +11,7 @@
 # limitations under the License.
 
 """
-Encapsulation of a single server definition.
+A single server definition including secrets.
 """
 
 from __future__ import absolute_import, print_function
@@ -21,7 +21,8 @@ __all__ = ['ServerDefinition']
 
 class ServerDefinition(object):
     """
-    Represents a single server definition from a server definition file.
+    Represents a single server definition from a server definition file,
+    and optionally the corresponding secrets from a vault file.
 
     Example for a server definition item in a server definition file:
 
@@ -33,23 +34,52 @@ class ServerDefinition(object):
           access_via: "VPN to dev network"
           user_defined:                         # user-defined part
             stuff: morestuff
+
+    Example for a corresponding secrets item in a vault file:
+
+    .. code-block:: yaml
+
+        myserver1:                              # nickname of the server
+          host: "10.11.12.13"
+          username: myuser1
+          password: mypass1
     """
 
-    def __init__(self, nickname, server_dict):
+    def __init__(self, nickname, server_dict, secrets_dict=None):
+        """
+        Parameters:
+
+          nickname (:term:`unicode string`):
+            Nickname of the server.
+
+          server_dict (dict):
+            Dictionary with the properties of the server item for the
+            nickname in the server definition file, with unspecified
+            properties defaulted.
+
+          secrets_dict (dict):
+            Dictionary with the properties of the secrets item for the
+            nickname in the vault file, or `None` if no vault file is specified
+            in the server definition file, or if the vault file does not
+            contain an entry for the server nickname.
+        """
         self._nickname = nickname
         self._description = server_dict['description']
         self._contact_name = server_dict.get('contact_name', None)
         self._access_via = server_dict.get('access_via', None)
         self._user_defined = server_dict.get('user_defined', None)
+        self._secrets = secrets_dict
 
     def __repr__(self):
+        secrets = "{...}" if self._secrets else "None"
         return "ServerDefinition(" \
             "nickname={s._nickname!r}, " \
             "description={s._description!r}, " \
             "contact_name={s._contact_name!r}, " \
             "access_via={s._access_via!r}, " \
-            "user_defined={s._user_defined!r})". \
-            format(s=self)
+            "user_defined={s._user_defined!r}, " \
+            "secrets={secrets})". \
+            format(s=self, secrets=secrets)
 
     @property
     def nickname(self):
@@ -101,3 +131,12 @@ class ServerDefinition(object):
         user-defined structure.
         """
         return self._user_defined
+
+    @property
+    def secrets(self):
+        """
+        dict: Secrets defined in the vault file for the server, or `None` if
+        no vault file is specified in the server definition file, or if the
+        vault file does not contain an entry for the server nickname.
+        """
+        return self._secrets